+++ arrays/index.js
@@ -1,0 +1,3 @@
+const sortByBigInt = require('./sort_by_big_int');
+
+module.exports = {sortByBigInt};