+++ index.js
@@ -1,4 +1,15 @@
 const {decodeBigSize} = require('./big_size');
+const {decodeTlvRecord} = require('./tlv_record');
+const {decodeTlvStream} = require('./tlv_stream');
 const {encodeBigSize} = require('./big_size');
+const {encodeTlvRecord} = require('./tlv_record');
+const {encodeTlvStream} = require('./tlv_stream');
 
-module.exports = {decodeBigSize, encodeBigSize};
+module.exports = {
+  decodeBigSize,
+  decodeTlvRecord,
+  decodeTlvStream,
+  encodeBigSize,
+  encodeTlvRecord,
+  encodeTlvStream,
+};