+++ tlv_record/index.js
@@ -1,0 +1,4 @@
+const decodeTlvRecord = require('./decode_tlv_record');
+const encodeTlvRecord = require('./encode_tlv_record');
+
+module.exports = {decodeTlvRecord, encodeTlvRecord};