+++ tlv_stream/index.js
@@ -1,0 +1,4 @@
+const decodeTlvStream = require('./decode_tlv_stream');
+const encodeTlvStream = require('./encode_tlv_stream');
+
+module.exports = {decodeTlvStream, encodeTlvStream};