+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Versions
 
+## 1.2.0
+
+- `decodeTlvRecord`: Decode a TLV record into a record
+- `decodeTlvStream`: Decode a TLV stream into component records
+- `encodeTlvRecord`: Encode record data as a TLV record
+- `encodeTlvStream`: Encode type value records as a TLV records stream
+
 ## 1.1.0
 
 - `decodeBigSize`: Add `length` attribute to indicate the BigSize byte length